Fix packing of union subtypes in Condition struct
The "cond_char" BitSet within the union is 32 bytes, which should be the
upper limit of the union size.

This commit fixes the size of the `cond_bufis` and `cond_str' fields to
be exactly 32 bytes in total. It also slightly increases the number of
bytes available by using uint8_t instead of size_t to store lengths.

Besides the minor improvements noted above, the main reason for these
changes is to replace incorrect and confusing expressions with simple
and obvious constants.

+++ src/syntax/highlight.c
@@ -23,7 +23,7 @@ static bool states_equal(void **ptrs, int idx, const State *b)
 
 static bool is_buffered(const Condition *cond, const char *str, size_t len)
 {
-    if (len != cond->u.cond_bufis.len) {
+    if (len != (size_t)cond->u.cond_bufis.len) {
         return false;
     }
